音楽室: unlock関連部分を切り出し
