+++ Maxence.vcxproj.filters
@@ -193,6 +193,9 @@
     <ClInclude Include="lib\modes\music_room.h">
       <Filter>ヘッダー ファイル</Filter>
     </ClInclude>
+    <ClInclude Include="lib\music_unlocker.h">
+      <Filter>ヘッダー ファイル</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <Image Include="Maxence.ico">
